Added support for 64 bit rates in the port_mapping isolator.
Currenty byte rates in the port_mapping isolator are represented by
uint32_t as in libnl 3.3. Because of that the maximum byte rate
supported by the isolator is ~34 Gbit/s (4294967295*8). This limits the
isolator's capability to utilize higher-speed NIC, e.g. 100G.

This patch changes byte rates to uint64_t, which allows us to support
support values up to ~18 Ebit/s (not UINT64_MAX because of the need to
convert bytes to bits for tc in the isolation script).

This change requires support from libnl. We have to bump the minimal
required version to 3.5.0 to get HTB API with 64 bit rate/ceil.

+++ src/linux/routing/queueing/htb.hpp
@@ -81,8 +81,8 @@ namespace cls {
 struct Config
 {
   Config(
-      uint32_t _rate,
-      Option<uint32_t> _ceil = None(),
+      uint64_t _rate,
+      Option<uint64_t> _ceil = None(),
       Option<uint32_t> _burst = None())
     : rate(_rate),
       ceil(_ceil),
@@ -99,9 +99,9 @@ struct Config
 
   // Normal rate limit. The size of the normal rate bucket is not
   // exposed and will be computed by the kernel.
-  uint32_t rate;
+  uint64_t rate;
   // Burst limit.
-  Option<uint32_t> ceil;
+  Option<uint64_t> ceil;
   // Size of the burst bucket.
   Option<uint32_t> burst;
 };
